support for EXEC CICS, EXEC SQL, EXEC SQLIMS, COMMENTRY ENTRY text; refactored preprocessor
