+++ cobol85/java/sub/CobolSubPreprocessor.java
@@ -1,0 +1,17 @@
+/*
+ * Copyright (C) 2017, Ulrich Wolffgang <u.wol@wwu.de>
+ * All rights reserved.
+ *
+ * This software may be modified and distributed under the terms
+ * of the BSD 3-clause license. See the LICENSE file for details.
+ */
+
+package io.proleap.cobol.preprocessor.sub;
+
+import io.proleap.cobol.preprocessor.CobolPreprocessor.CobolDialect;
+import io.proleap.cobol.preprocessor.CobolPreprocessor.CobolSourceFormatEnum;
+
+public interface CobolSubPreprocessor {
+
+	String processLines(String lines, CobolSourceFormatEnum format, CobolDialect dialect);
+}
