+++ cobol85/java/sub/impl/ThrowingErrorListener.java
@@ -1,0 +1,22 @@
+/*
+ * Copyright (C) 2017, Ulrich Wolffgang <u.wol@wwu.de>
+ * All rights reserved.
+ *
+ * This software may be modified and distributed under the terms
+ * of the BSD 3-clause license. See the LICENSE file for details.
+ */
+
+package io.proleap.cobol.preprocessor.sub.impl;
+
+import org.antlr.v4.runtime.BaseErrorListener;
+import org.antlr.v4.runtime.RecognitionException;
+import org.antlr.v4.runtime.Recognizer;
+
+public class ThrowingErrorListener extends BaseErrorListener {
+
+	@Override
+	public void syntaxError(final Recognizer<?, ?> recognizer, final Object offendingSymbol, final int line,
+			final int charPositionInLine, final String msg, final RecognitionException e) {
+		throw new RuntimeException("syntax error in line " + line + ":" + charPositionInLine + " " + msg);
+	}
+}
