+++ cobol85/java/sub/parser/CobolParserPreprocessor.java
@@ -1,0 +1,15 @@
+/*
+ * Copyright (C) 2017, Ulrich Wolffgang <u.wol@wwu.de>
+ * All rights reserved.
+ *
+ * This software may be modified and distributed under the terms
+ * of the BSD 3-clause license. See the LICENSE file for details.
+ */
+
+package io.proleap.cobol.preprocessor.sub.parser;
+
+import io.proleap.cobol.preprocessor.sub.CobolSubPreprocessor;
+
+public interface CobolParserPreprocessor extends CobolSubPreprocessor {
+
+}
